Контроллер для сотрудников
