+++ WebStore_GB/Views/Shared/Partial/_MainMenu.cshtml
@@ -17,6 +17,7 @@
             <li><a href="blog-single.html">Blog Single</a></li>
         </ul>
     </li>
+    <li><a asp-controller="Employees" asp-action="Index">Сотрудники</a></li>
     <li><a href="404.html">Error page 404</a></li>
     <li><a href="contact-us.html">Contact</a></li>
 </ul>
